+++ cmake/UseRPMTools.cmake
@@ -168,6 +168,7 @@ rm -rf build_tree
 	COMMAND cpack3 -G TGZ --config CPackSourceConfig.cmake
 	COMMAND ${CMAKE_COMMAND} -E copy ${CPACK_SOURCE_PACKAGE_FILE_NAME}.tar.gz ${RPM_ROOTDIR}/SOURCES    
 	COMMAND ${RPMTools_RPMBUILD_EXECUTABLE} -bb --define=\"_topdir ${RPM_ROOTDIR}\" $ENV{RPMDEFS} --buildroot=${RPM_ROOTDIR}/tmp ${RPM_ROOTDIR}/SPECS/${SPECFILE_NAME} 
+  JOB_POOL console
 	)  
     ENDMACRO(RPMTools_ADD_RPM_TARGETS)