+++ continuousintegration/ci_helpers/build_srpm.sh
@@ -22,6 +22,7 @@ usage() {
   echo ""
   echo "Builds the srpms."
   echo "  --build-dir <build-dir>:              Sets the build directory for the SRPMs. Can be absolute or relative to where the script is being executed from."
+  echo "  --build-generator <generator>:        Specifies the build generator for cmake. Ex: [\"Unix Makefiles\", \"Ninja\"]."
   echo "  --scheduler-type <scheduler-type>:    The scheduler type. Ex: objectstore."
   echo "  --cta-version <cta-version>:          Sets the CTA_VERSION."
   echo "  --vcs-version <vcs-version>:          Sets the VCS_VERSION variable in cmake."
@@ -30,7 +31,8 @@ usage() {
   echo "options:"
   echo "  -i, --install:                        Installs the required packages. Supported operating systems: [cc7, alma9]."
   echo "  -j, --jobs <num-jobs>:                How many jobs to use for make."
-  echo "      --create-build-dir                        Creates the build directory if it does not exist."
+  echo "      --clean-build-dir:                Empties the build directory, ensuring a fresh build from scratch."
+  echo "      --create-build-dir                Creates the build directory if it does not exist."
   echo "      --skip-unit-tests:                Skips the unit tests."
   echo "      --oracle-support <ON/OFF>:        When set to OFF, will disable Oracle support. Oracle support is enabled by default."
   echo "      --cmake-build-type <build-type>:  Specifies the build type for cmake. Must be one of [Release, Debug, RelWithDebInfo, or MinSizeRel]."
@@ -45,12 +47,14 @@ build_srpm() {
 
   # Default values for arguments
   local build_dir=""
+  local build_generator=""
   local cta_version=""
   local scheduler_type=""
   local vcs_version=""
   local xrootd_version=""
 
   local create_build_dir=false
+  local clean_build_dir=false
   local install=false
   local num_jobs=1
   local skip_unit_tests=false
@@ -70,6 +74,19 @@ build_srpm() {
           usage
         fi
         ;;
+      --build-generator) 
+        if [[ $# -gt 1 ]]; then
+          if [ "$2" != "Ninja" ] && [ "$2" != "Unix Makefiles" ]; then
+              echo "Warning: build generator $2 is not officially supported. Compilation might not be successful."
+          fi
+          build_generator="$2"
+          shift
+        else
+          echo "Error: --build-generator requires an argument"
+          usage
+        fi
+        ;;
+      --clean-build-dir) clean_build_dir=true ;;
       --create-build-dir) create_build_dir=true ;;
       --cta-version)
         if [[ $# -gt 1 ]]; then
@@ -167,6 +184,11 @@ build_srpm() {
     usage
   fi
 
+  if [ -z "${build_generator}" ]; then
+    echo "Failure: Missing mandatory argument --build-generator";
+    usage
+  fi
+
   if ! xrootd_supported "${xrootd_version}"; then
     echo "Unsupported xrootd-version: ${xrootd_version}. Must be one of [4, 5]."
     exit 1
@@ -177,6 +199,11 @@ build_srpm() {
   cd ../../
   local repo_root=$(pwd)
   local cmake_options=""
+
+  if [[ ${clean_build_dir} = true ]]; then
+    echo "Removing old build directory: ${build_dir}"
+    rm -rf "${build_dir}"
+  fi
 
   if [[ ${create_build_dir} = true ]]; then
     mkdir -p "${build_dir}"
@@ -200,11 +227,28 @@ build_srpm() {
       cp -f continuousintegration/docker/ctafrontend/alma9/etc/yum.repos.d/*.repo /etc/yum.repos.d/
       cp -f continuousintegration/docker/ctafrontend/alma9/etc/yum/pluginconf.d/versionlock.list /etc/yum/pluginconf.d/
       yum install -y epel-release almalinux-release-devel
-      yum install -y wget gcc gcc-c++ cmake3 make rpm-build yum-utils
+      yum install -y wget gcc gcc-c++ cmake3 rpm-build yum-utils
+      case "${build_generator}" in
+        "Unix Makefiles")
+          yum install -y make
+          ;;
+        "Ninja")
+          yum install -y ninja-build
+          ;;
+        *)
+          echo "Failure: Unsupported build generator for alma9: ${build_generator}"
+          exit 1
+          ;;
+      esac
       ./continuousintegration/docker/ctafrontend/alma9/installOracle21.sh
     elif [ "$(grep -c 'CentOS Linux release 7' /etc/redhat-release)" -eq 1 ]; then
       # CentOS 7
       echo "Found CentOS 7 install..."
+      if [[ ! ${build_generator} = "Unix Makefiles" ]]; then
+        # We only support Unix Makefiles for cc7
+        echo "Failure: Unsupported build generator for cc7: ${build_generator}"
+        exit 1
+      fi
       cp -f continuousintegration/docker/ctafrontend/cc7/etc/yum.repos.d/*.repo /etc/yum.repos.d/
       if [[ ${xrootd_version} -eq 4 ]]; then
         echo "Using XRootD version 4"
@@ -219,6 +263,7 @@ build_srpm() {
       source /opt/rh/devtoolset-11/enable
     else
       echo "Failure: Unsupported distribution. Must be one of: [cc7, alma9]"
+      exit 1
     fi
   fi
 
@@ -250,11 +295,11 @@ build_srpm() {
 
   cd "${build_dir}"
   echo "Executing cmake..."
-  cmake3 ${cmake_options} "${repo_root}"
+  (set -x; cmake3 ${cmake_options} -G "${build_generator}" "${repo_root}")
 
-  # Make
-  echo "Executing make..."
-  make cta_srpm -j "${num_jobs}"
+  # Build step
+  echo "Executing build step using: ${build_generator}"
+  cmake --build . --target cta_srpm -- -j "${num_jobs}"
 }
 
 build_srpm "$@"